// src/swagger/model/vehicleDetailsOffRoad.ts
/**
 * fms
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface VehicleDetailsOffRoad { 
    zone?: string;
    region?: string;
    branch?: string;
    registration_No?: string;
    model?: string;
    make_Year?: string;
    offroad_Reason?: string;
    offroad_From_Date?: string;
    is_vehicle_offroad?: string;
    offRoad_Age?: number;
    totalVehicles?: number;
}